Big & (Hopefully) Final ECS Rewrite
Yes, another one. Another rewrite. Currently, most functionality is untested. Most of the important stuff is implemented. So what has changed? Well, 'managers' are now planets, and systems don't really exist anymore. ArchetypeManager is now split into many little planets, as it should be. Entities have generational indices meaning that they can be stored and used for component access later. The big, big change is that the whole concept of maybe using ffi is out of the window. We don't need it. The two big problems now are logging, deadlocks, and testing, so I'll be back when I've hopefully sorted those out.