Macos mewo_gpu support?
Well, I got a macos kvm thinking I'd be able to test it. Well, I got it
compiling and fixed an issue with instance extensions, but it turns out
that the kvm had no gpu and therefore could not run metal, needed for
vulkan. So yes? Maybe? All I can say is that it *should* work on macos.